+++ native_extension/ane/build_ios.sh
@@ -23,6 +23,9 @@ fi
 if [ ! -d "$pathtome/platforms/ios/simulator" ]; then
 mkdir "$pathtome/platforms/ios/simulator"
 fi
+if [ ! -d "$pathtome/platforms/ios/device" ]; then
+mkdir "$pathtome/platforms/ios/device"
+fi
 if [ ! -d "$pathtome/platforms/ios/default" ]; then
 mkdir "$pathtome/platforms/ios/default"
 fi
@@ -39,12 +42,14 @@ unzip "$pathtome/$PROJECT_NAME.swc" "library.swf" -d "$pathtome"
 #Copy library.swf to folders.
 echo "Copying library.swf into place."
 cp "$pathtome/library.swf" "$pathtome/platforms/ios/simulator"
+cp "$pathtome/library.swf" "$pathtome/platforms/ios/device"
 cp "$pathtome/library.swf" "$pathtome/platforms/ios/default"
 
 #Copy native libraries into place.
 echo "Copying native libraries into place."
 cp -R -L "$pathtome/../../native_library/ios/$PROJECT_NAME/Build/Products/Release-iphonesimulator/lib$PROJECT_NAME.a" "$pathtome/platforms/ios/simulator"
 
+cp -R -L "$pathtome/../../native_library/ios/$PROJECT_NAME/Build/Products/Release-iphoneos/lib$PROJECT_NAME.a" "$pathtome/platforms/ios/device"
 
 #Run the build command.
 echo "Building Simulator Release."
@@ -53,10 +58,13 @@ echo "Building Simulator Release."
 -swc "$pathtome/$PROJECT_NAME.swc" \
 -platform iPhone-x86  -C "$pathtome/platforms/ios/simulator" "library.swf" "lib$PROJECT_NAME.a" \
 -platformoptions "$pathtome/platforms/ios/platform.xml" \
+-platform iPhone-ARM  -C "$pathtome/platforms/ios/device" "library.swf" "lib$PROJECT_NAME.a" \
+-platformoptions "$pathtome/platforms/ios/platform.xml" \
 -platform default -C "$pathtome/platforms/ios/default" library.swf
 
 
 rm -r "$pathtome/platforms/ios/simulator"
+rm -r "$pathtome/platforms/ios/device"
 rm -r "$pathtome/platforms/ios/default"
 rm "$pathtome/$PROJECT_NAME.swc"
 rm "$pathtome/library.swf"