+++ native_library/source/ANEhelper.cpp
@@ -2,11 +2,22 @@
 #include "FlashRuntimeExtensions.h"
 #include <windows.h>
 #include <conio.h>
-#else
+#elif __APPLE__
+#include "TargetConditionals.h"
+#if TARGET_IPHONE_SIMULATOR || TARGET_OS_IPHONE
+#include "FlashRuntimeExtensions.h"
+#elif TARGET_OS_MAC
 #include <Adobe AIR/Adobe AIR.h>
 #include <stdlib.h>
 #include <stdio.h>
+#else
+#   error "Unknown Apple platform"
 #endif
+
+#endif
+
+
+
 #include "../include/ANEhelper.h"
 #include <string>
 #include <vector>